A few changes
Fixed a few bugs regarding player name changing after successful login. PacketDirection is now integrated into the packet library itself. PlayerConnection instance is now synced on ServerToClient and ClientToServer handlers. I will need to work on collecting a list of connected players next. Chat message parsing is now possible, which should allow for proxy commands. The API for that functionality is not yet completed in XTerraPacket, however, it is usable.